+++ data/definitions.ts
@@ -40,7 +40,7 @@ export interface Proxy {
 }
 
 export interface EvmCoreContracts {
-  deployHeight: number,
+  deployHeight: number;
   upgradeBeaconController: NomadIdentifier;
   xAppConnectionManager: NomadIdentifier;
   updaterManager: NomadIdentifier;
@@ -61,7 +61,7 @@ export interface DeployedCustomToken {
 }
 
 export interface EvmBridgeContracts {
-  deployHeight: number,
+  deployHeight: number;
   bridgeRouter: Proxy;
   tokenRegistry: Proxy;
   bridgeToken: Proxy;
@@ -107,6 +107,8 @@ export interface CustomTokenSpecifier {
 export interface BridgeConfiguration {
   weth?: NomadIdentifier;
   customs: Array<CustomTokenSpecifier>;
+  mintGas: number | string;
+  deployGas: number | string;
 }
 
 export interface Domain {